NSFontManager: clean up -convertFont:toHaveTrait:
Try not to call -fontWithFamily:traits:weight:size: in -convertFont:*,
because convertFont: methods are only supposed to change exactly one
thing about a font (and they're allowed to fail, returning the original
font).

I noticed that the text system was converting one font into another just
because the original didn't have an italic face. This is all kinds of wrong,
especially because when the text system does that it gets saved out to disk.

(that's another thing that needs to get fixed -- substituting glyphs is OK
but it ought to be temporary, leaving no trace in files that it occurred)