Standard ruler should not be in points...
Set up default ruler view for inches, in the future we can make it
configurable but we should at the very least make it useful for somebody.

While we're at it, when we're setting up the default tabs, let's create them
in half-inch increments.

There is a reasonable technical rationale for this: a half inch is 36
points, which is equal to 36 pixels at the 72 dpi we work with internally,
and we can easily convert whole numbers to weird output measurements because
the numbers work out in unscaled space well.

In comparison, a centimeter is not 28 points, it's ~28.34645669291338582677,
so going back and re-scaling everything gets complicated in a hurry.
Centimeters only have roughly normal numbers of pixels at multiples of
127 dpi, and half-pixels at multiples of 63.5 dpi.

You can mostly get away with rounding down to 28, but over a 508mm
screen width the spacing difference between 28 pts. and 1 cm grows to
~17.6 points, or 0.6208... of a cm. Which is probably fine for many cases,
but you probably want to know for sure.

Either way, doing the ruler with hundred-point markings and with twelve
28-point tabs seems silly.